# perf(kit): optimize Nuxt module resolving