### 🔗 Linked issue

<!-- Please ensure there is an open issue and mention its number. For example, "resolves #123" -->


### 📚 Description

When Nuxt loads modules using the `modules` config key in `nuxt.config`, it's quite forgiving. It resolves aliases, and relative paths and will check subpath exports of `/nuxt` and `/module` from the string path.

This loading strategy is good DX but is not optimized for how most Nuxt modules are published and loaded.

The [nuxt-module-builder](https://github.com/nuxt/module-builder) does not suggest subpath exports, instead, it only adds a main entry. Meaning that modules published the official way, are being loaded inefficiently as subpath exports are being checked first.

This results in an excessive waste of time trying to resolve paths that don't exist.

To solve this is quite simple, we just reorder the path resolve order. 

:question: We may also consider checking if the `src` exists before trying to import it via `jiti` as a further optimization.

:fire: Benchmarks

https://github.com/harlan-zw/unhead.unjs.io

Across 5 runs, an average of 4,961.4 ms was wasted trying to resolve non-existent module paths. Switching the path order in this PR wasted 0ms.

This resulted in:
- `nuxi prepare` 8 seconds -> 3 seconds (62% faster)
- `nuxi dev` 20 seconds -> 15 seconds (25% faster)

<!-- Describe your changes in detail. Why is this change required? What problem does it solve? -->

<!----------------------------------------------------------------------
Before creating the pull request, please make sure you do the following:

- Check that there isn't already a PR that solves the problem the same way. If you find a duplicate, please help us reviewing it.
- Read the contribution docs at https://nuxt.com/docs/community/contribution
- Ensure that PR title follows conventional commits (https://www.conventionalcommits.org)
- Update the corresponding documentation if needed.
- Include relevant tests that fail without this PR but pass with it.

Thank you for contributing to Nuxt!
----------------------------------------------------------------------->